Downgrade Cairo to make it work with 7

+++ gfx/cairo/cairo/src/cairo-cache-private.h
@@ -84,7 +84,7 @@
  * not be initialized if so desired.
  **/
 typedef struct _cairo_cache_entry {
-    uintptr_t hash;
+    unsigned long hash;
     unsigned long size;
 } cairo_cache_entry_t;
